Each individual event class appends a class name.
Each event includes its class name, or some other's class class name
(if it is trying to appear as that event).

We used to append that class name in one central place, but that
prevented each event from fine-tuning the seed that is goning to be
used for digest calculation down the line.

Now each class must, or at least should (to prevent false duplicates)
include its class name. If not its own, than it should provide the class
name of the the event type it's trying to mimic, e.g. FakeFollowEvent
mimics a real FollowEvent.


Former-commit-id: a40a44edf6346d0e5c0fc5c46f11fbc59e95a250