Removed FakeDigestCreator and all refs to it.
The use case it was designed for is no longer valid. We don't grab stuff
manually and store it to ApiCache any more. This was used for filling
missing follows/stars/watches. Bithub PaaS doesn't support that. Even if
we at one point decide to support it, it should be automatic, not
manual, so we'll need to reimplement it properly.