Prototype of the new system for filtering.
Each embed has a blocking and a moderating filter.
Each filter has one or more (natural language) queries.
Each query is stored as a sentence that is translated into AR query at
qury-time.

Queries are stored as sentences because of the flexibility that
provides. We can change the meaning of the subjects, verbs and objects
of the query and the only thing we need to change is the
QueryTranslator.