Payment model.
Listens for stripe's 'invoice.payment_succeeded' webhook.

When new invoice is created and payment is successful, stripe sends
notif via webhook. Invoice data (plan, date, ids, amount) are saved as
Payment.

Payment should be never updated, it kind-of copy of stripe's log so that
we can display billing history to the brand manager.


Former-commit-id: 02fbb5849702434d8bb8318f36d241862125e477