Elevated a few data structures to classes.
TreePath represents the actors position in the supervision tree. Used
for naming actors and keeping the context and the position of the actor
relative to the its parent. It has methods for calculating the current
path and finding various information in the tree.

ServiceInfo represents the combined service feed and service type.


Former-commit-id: 0b38e1db78e0168b85fff5a8d683d71564e206eb