Analytics.
A new histogram table/model was added that stores historic data about
embeds/services where each row has a **volume** (count) of entities
in the specified source (embed/service), the **delta** compared to the
last measurement (row) and the timestamp the values were measured at.

A new worker has been added that fills this table regularly and
a controller that exposes the data in various ways.


Former-commit-id: 46519737267c9cb7a4b11b137fa2a9af1dcafb2b