Propagation reworked.
Commands are now routed through the supervision tree. In AMQP terms,
"direct" instead of "fanout". That makes the "i_am_targets_parent?"
check unnecessary.

Each Supervisor now only handles the comman if the message is intended
for that exact level. For example, a BrandInfo command/node is only
handled in Supervisor::Main (which is a parent of Supervisor::Brand).


Former-commit-id: afc565665e026369ccbf109a830f2ab177f5fde8