When deleting services or embeds, clear links too.
1.
When deleting an embed, delete all entities that became orphaned
(belong to no embed).

2.
When deleting a service, delete connections between entities
belonging to the service we're currently deleting and the embed
that service belongs to.

Afterwards, delete connections between entities and the service we're
currently deleting.

Finally, delete all entities that belong to no service (orphaned
entities).


Former-commit-id: 87052a589f3d8050a00ed4a6d9b86745a49a7b32 [formerly cfb90e1b3095515a92249b720814c574168d4959]
Former-commit-id: 7b400a474a8a8c9779889d8f9ea33f23d4fdbf0e