When deleting services or embeds, clear links too.
1.
When deleting an embed, delete all entities that became orphaned
(belong to no embed).

2.
When deleting a service, delete connections between entities
belonging to the service we're currently deleting and the embed
that service belongs to.

Afterwards, delete connections between entities and the service we're
currently deleting.

Finally, delete all entities that belong to no service (orphaned
entities).