Fixed case where manual moderation should overwrite automatic moderation.
Manual moderation column (is_approved_manually) in embed_entities can be
null. If it is, that means that the user has not made a decision.
In that case, read the value from the automatic moderation column
(is_approved_automatically).

Also, always read the is_approved method in Entity and EmbedEntity that
picks the right column to read from.