Fixed case where manual moderation should overwrite automatic moderation.
Manual moderation column (is_approved_manually) in embed_entities can be
null. If it is, that means that the user has not made a decision.
In that case, read the value from the automatic moderation column
(is_approved_automatically).

Also, always read the is_approved method in Entity and EmbedEntity that
picks the right column to read from.


Former-commit-id: 4fe5c74972c64787df30a10a4c8762494570bbea [formerly a7f92e20ab415ba08844c7e09f443311467d2218] [formerly eea0b19d25bee44e0834e38086018bd9a9325a7a [formerly 12d4acb970eea949ae5cf95ad1e8cbab63a6b3ea]]
Former-commit-id: 8ba761581a4bffd47b3f7770c253c07e5a86ff28 [formerly a32438d8ae99061a2ed56981b547d8bb4d729ce7]
Former-commit-id: 8e7a1e3bf02abc014ec1f65be0a9ebe13407e3fc