Test for digest set membership in batches.
We used to test for digest set membership event-by-event and that
caused the publisher to bloat in memory because waiting IO methods
kept piling up in it's mailbox.

We now fetch the whole set and test for membership in memory.

This way we perform one IO operation per batch instead
N operations, where N is the size of the batch.