Test for digest set membership in batches.
We used to test for digest set membership event-by-event and that
caused the publisher to bloat in memory because waiting IO methods
kept piling up in it's mailbox.

We now fetch the whole set and test for membership in memory.

This way we perform one IO operation per batch instead
N operations, where N is the size of the batch.


Former-commit-id: e31a1da1d7b6ffbf3cd3be9a73201b92e5a6e899 [formerly c589802c8f746bd7b61835a4f68e2132e603bbb8] [formerly de48dda4f2af8a0a0b941031c605b77d5e79eef2 [formerly ecfea556c29646cf643f63f0820004eb84d087d9]] [formerly 7ad2f93143510dfa2b7c52addf28e51eadca1767 [formerly a068f121fcface887e6f2887cbd9901fd7b35d92] [formerly 5d0b2e13a0e25e19afa57d6bba6143e04bccfd94 [formerly 00eaf66012470363478f7b4258d79bad1736d6b1]]]
Former-commit-id: 857c0abac11021009e569c07c3c9e7bdc8ef9762 [formerly 4494db196b26b5835d833fc5a69932a218e2fa02] [formerly 90eb3697db078fdbd4971399301084084f1ad46a [formerly 5fed06026a5d3561dfe92392fbd31ebd61d261e1]]
Former-commit-id: 9c56350b351e9b9774601bf7d05d4046a9b4d29d [formerly 77327012abc09b1b45d6a165e11fdb45685d3ffd]
Former-commit-id: 28b990771a8251442e417c4d3755848d6f01bdc8